remove test/avio.jl
